+++ acceleration/configuration.py
